+++ CHANGELOG.md
@@ -53,7 +53,7 @@ Items that are related, such as breaking changes, new features, or changes to ex
 
 ### Changed
 
-- [@siteimprove/alfa-rules](packages/alfa-rules): SIA-R65 now automatically accepts difference in `color` as a valid focus indicator. ([#658](../../issues/658), [#713](../../pull/713))
+- [@siteimprove/alfa-rules](packages/alfa-rules): SIA-R65 now automatically accepts difference in `background-color` and `color` as a valid focus indicator. ([#658](../../issues/658), [#713](../../pull/713), [#714](../../pull/714))
 
 ### Fixed